+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/GameRendererMixin.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.polytone.mixins;
 
+import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.lightmap.LightmapsManager;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.LightTexture;
@@ -23,14 +24,16 @@ public abstract class GameRendererMixin {
     @Inject(method = "render", at = @At(value = "NEW",
             target = "(Lnet/minecraft/client/Minecraft;Lnet/minecraft/client/renderer/MultiBufferSource$BufferSource;)Lnet/minecraft/client/gui/GuiGraphics;"))
     private void polytone$messWithGui(float partialTicks, long nanoTime, boolean renderLevel, CallbackInfo ci) {
-        LightmapsManager.setupForGUI(true);
+        Polytone.LIGHTMAPS.setupForGUI(true);
         lightTexture.turnOnLightLayer();
+        Polytone.OVERLAY_MODIFIERS.onStartRenderingOverlay();
     }
 
     @Inject(method = "render", at = @At(value = "TAIL"))
     private void polytone$resetGuiLightmap(float partialTicks, long nanoTime, boolean renderLevel, CallbackInfo ci) {
-        LightmapsManager.setupForGUI(false);
+        Polytone.LIGHTMAPS.setupForGUI(false);
         lightTexture.turnOnLightLayer();
+        Polytone.OVERLAY_MODIFIERS.onStartRenderingOverlay();
     }
 
 }