+++ changelog.md
@@ -1,1 +1,1 @@
-fixed a packge issue
+- added overlay modifiers, a new tool to modify screen overlay sprites.Change their size, position and UVs